refactor: simplify AuthContext to only hold isAuthenticated state
